GraphController
