+++ WebApp/Controllers/AuthController.cs
@@ -104,104 +104,6 @@ namespace WebApp.Controllers
             });
         }
 
-        [HttpPost("graph")]
-        public IActionResult CreateGraph(CreateGraphDto dto)
-        {
-            User user;
-            try
-            {
-                string jwt = Request.Cookies["jwt"];
 
-                JwtSecurityToken token = _jwtService.Verify(jwt);
-
-                int userId = int.Parse(token.Issuer);
-
-                user = _repository.GetById(userId);
-            }
-            catch (Exception)
-            {
-                return Unauthorized();
-            }
-
-            GraphModel graph = new GraphModel
-            {
-                Name = dto.Name,
-                Data = dto.Data,
-            };
-
-            try
-            {
-                _repository.CreateGraph(user, graph);
-            }
-            catch (Exception)
-            {
-                return BadRequest();
-            }
-
-            return Created("success", graph);
-        }
-
-        [HttpGet("graph")]
-        public IActionResult ReadGraph(string name)
-        {
-            User user;
-
-            try
-            {
-                string jwt = Request.Cookies["jwt"];
-
-                JwtSecurityToken token = _jwtService.Verify(jwt);
-
-                int userId = int.Parse(token.Issuer);
-
-                user = _repository.GetById(userId);
-
-            }
-            catch (Exception)
-            {
-                return Unauthorized();
-            }
-
-            GraphModel graph = _repository.ReadGraph(name, user);
-
-            if (graph == default(GraphModel))
-            {
-                return NotFound();
-            }
-            else
-            {
-                return Ok(graph);
-            }
-        }
-
-
-        [HttpPut("graph")]
-        public IActionResult UpdateGraph(CreateGraphDto dto)
-        {
-            User user;
-
-            try
-            {
-                string jwt = Request.Cookies["jwt"];
-                JwtSecurityToken token = _jwtService.Verify(jwt);
-                int userId = int.Parse(token.Issuer);
-                user = _repository.GetById(userId);
-            }
-            catch (Exception)
-            {
-                return Unauthorized();
-            }
-
-            GraphModel graph = _repository.UpdateGraph(user, dto.Name, dto.Data);
-
-            if (graph == default(GraphModel))
-            {
-                return NotFound();
-            }
-            else
-            {
-                return Ok(graph);
-            }
-        }
     }
 }
