+++ WebApp/Data/IUserRepository.cs
@@ -10,11 +10,7 @@ namespace WebApp.Data
 
         User GetById(int id);
 
-        GraphModel CreateGraph(User user, GraphModel graph);
 
-        GraphModel ReadGraph(string graphName, User user);
-
-        GraphModel UpdateGraph(User user, string graphName, string newData);
 
     }
 }
